refact: ディレクトリ構成など変更
