+++ package.json
@@ -24,7 +24,14 @@
     "dotenv": "^16.3.1",
     "form-data": "^4.0.0",
     "fs": "^0.0.1-security",
+    "module-alias": "^2.2.3",
     "path": "^0.12.7",
     "selenium-webdriver": "^4.16.0"
+  },
+  "_moduleAliases": {
+    "@root": ".",
+    "@src": "./src",
+    "@slack-app": "./src/slack-app",
+    "@utils": "./src/utils"
   }
 }
