+++ src/index.js
@@ -1,4 +1,6 @@
-const SlackApp = require("./slack-app");
+// package.jsonに設定したパスエイリアスを登録
+require("module-alias/register");
 
+const SlackApp = require("./slack-app/slack-app");
 const slackApp = new SlackApp();
 slackApp.start();
