+++ src/utils/line-stamp-util.js
