+++ src/utils/slack-api-client/get-token-and-cookie.js
@@ -1,8 +1,5 @@
 const { Builder, By, Key, until } = require("selenium-webdriver");
 const chrome = require("selenium-webdriver/chrome");
-const fs = require("fs");
-const FormData = require("form-data");
-const { default: axios } = require("axios");
 
 // 環境変数の読み込み
 require("dotenv").config();
@@ -11,38 +8,8 @@ const email = process.env.EMAIL;
 const password = process.env.PASSWORD;
 
 /**
- * LINEスタンプをSlack絵文字として登録する関数
- * @param {string} emojiNamePrefix - 絵文字の登録名のプレフィックス
- * @param {string} stampsFilePathList - LINEスタンプの画像ファイルのパスのリスト
- * @returns {Promise<object[]>} - 絵文字の登録結果のリスト
- */
-async function registerAllLineStampsToSlack(
-  emojiNamePrefix,
-  stampsFilePathList
-) {
-  // POSTリクエストに必要なCookieとAPIトークンを取得
-  const { cookie, token } = await getCookieAndToken();
-
-  // 全ての絵文字の登録処理を並列で実行
-  const registrationPromises = stampsFilePathList.map((stampsFilePath, idx) => {
-    const emojiName = `${emojiNamePrefix}_${idx}`;
-
-    return registerOneLineStampToSlack(
-      emojiName,
-      stampsFilePath,
-      cookie,
-      token
-    );
-  });
-
-  // すべての絵文字の登録処理が終わるまで待機
-  const results = await Promise.all(registrationPromises);
-  console.log(results);
-  return results;
-}
-
-/**
  * Seleniumを使ってサインインしてCookieとAPIトークンを取得する関数.
+ * @returns {Promise<object>} - CookieとAPIトークン
  */
 async function getCookieAndToken() {
   // Seleniumのドライバーをビルド
@@ -94,7 +61,7 @@ async function signInToSlack(driver) {
   await driver.findElement(By.id("email")).sendKeys(email);
   await driver.findElement(By.id("password")).sendKeys(password, Key.RETURN);
 
-  // 'boot_data'を含むscriptタグが読みこまれるのを待機
+  // 'boot_data'を含むscriptタグが読みこまれるのを待機 (api-tokenを取得するのに必要)
   await driver.wait(
     until.elementLocated(By.xpath("//script[contains(text(), 'boot_data')]")),
     10000
@@ -118,49 +85,11 @@ async function getCookie(driver) {
 
 /**
  * APIトークンを取得する関数
- * @param {WebDriver} driver
+ * @param {WebDriver} driver - Seleniumのドライバー
  * @returns {Promise<string>} - APIトークン
  */
 async function getToken(driver) {
   return driver.executeScript("return window.boot_data.api_token;");
 }
 
-/**
- * 1つのLINEスタンプをSlack絵文字として登録する関数
- * @param {string} emojiName - 絵文字の登録名
- * @param {string} stampsFilePath - LINEスタンプの画像ファイルのパス
- * @param {string} cookie - Cookie
- * @param {string} token - APIトークン
- * @returns {Promise<object>} - 絵文字の登録結果
- */
-async function registerOneLineStampToSlack(
-  emojiName,
-  stampsFilePath,
-  cookie,
-  token
-) {
-  const url = `https://${workspace}.slack.com/api/emoji.add`;
-
-  // FormData作成
-  const formData = new FormData();
-  const file = fs.createReadStream(stampsFilePath);
-  formData.append("name", emojiName);
-  formData.append("image", file);
-  formData.append("token", token);
-  formData.append("mode", "data");
-
-  return axios
-    .post(url, formData, {
-      headers: {
-        Cookie: cookie,
-      },
-    })
-    .then((response) => {
-      return { name: emojiName, ...response.data };
-    })
-    .catch((error) => {
-      return { name: emojiName, error };
-    });
-}
-
-module.exports = registerAllLineStampsToSlack;
+module.exports = getCookieAndToken;
